chore(sdk): Move `RoomEventCache` in `caches/room/mod.rs`.
